Feat - Hospital Search - MapView
- 지도 보여주는 화면
- FIX : Api 클라이언트 설정
- FIX : Filter BottomSheet
- TODO : ViewModel 페이징 로직 필요

+++ data/src/main/java/com/example/data/datasource/remote/network/feature/hospital/hospital/dto/HospitalsResDto.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class HospitalsResDto(
-    val hospitalId: Int,
+    val hospitalId: Long,
     val name: String,
     val lat: Double,
     val lng: Double,
